hw: use a no-devices placeholder component
