+++ liana-ui/src/component/modal/mod.rs
@@ -6,7 +6,7 @@ use iced::{
         button::{Status, Style},
         column, row, Space,
     },
-    Length, Padding,
+    Alignment, Length, Padding,
 };
 
 use iced::widget::Container;
@@ -23,6 +23,8 @@ use crate::{
 };
 
 use crate::widget::{Button, CheckBox, Column, ColumnExt, Element, Row, RowExt, SpaceExt, Text};
+
+use super::text::p1_regular;
 
 pub const BTN_W: u32 = 500;
 pub const V_SPACING: u32 = 10;
@@ -430,4 +432,35 @@ where
 
     let msg = on_press.map(|f| f());
     button::device(col, msg)
+}
+
+pub fn show_qr_code<'a, M: 'a + 'static>(
+    tt: Option<&'static str>,
+    msg: Option<M>,
+) -> Button<'a, M> {
+    let mut btn = Button::new(
+        Row::new()
+            .push(icon::qr_icon().size(30))
+            .push(p1_regular("Show QR Code"))
+            .push_maybe(tt.map(super::tooltip))
+            .spacing(20)
+            .align_y(Alignment::Center)
+            .padding(10 + 5),
+    )
+    .style(theme::button::secondary)
+    .width(Length::Fill);
+    if let Some(msg) = msg {
+        btn = btn.on_press(msg);
+    }
+    btn
+}
+
+pub fn modal_no_devices_placeholder<'a, M: 'a>() -> Element<'a, M> {
+    column![
+        icon::usb_icon().size(100),
+        p1_regular("Plug in a hardware device ...")
+    ]
+    .align_x(Horizontal::Center)
+    .spacing(20)
+    .into()
 }
